Fix text field updates not applying when e.g. clicking into the codemirror editor
ARGHHHH

So there was a problem with like, empty string categories/etc. being sent to the server.
The thing was, the modify() ran for the text field, set the category to "something",
then instantly the modify() ran for the codemirror editor, changing the markdown text to "",
but it was working on an old snapshot of the entry, untouched by the category modification,
because there was no parent rerendering in between these two modify() calls.

The update() forces a rerender when we reassign the entry.
And getting rid of the cached-in-render() entry was also required o_0 aaaaaaa
(but it was kinda silly anyway?)